TUIKitDemo upgrade 8.0.5895

+++ iOS/TUIKit/TUIChat/UI_Classic/Input/TUIInputController.m
@@ -424,7 +424,7 @@
 - (void)inputBar:(TUIInputBar *)textView didSendVoice:(NSString *)path {
     NSURL *url = [NSURL fileURLWithPath:path];
     AVURLAsset *audioAsset = [AVURLAsset URLAssetWithURL:url options:nil];
-    int duration = CMTimeGetSeconds(audioAsset.duration);
+    float duration = (float)CMTimeGetSeconds(audioAsset.duration);
     int formatDuration = duration > 59 ? 60 : duration + 1 ;
     V2TIMMessage *message = [[V2TIMManager sharedInstance] createSoundMessage:path duration:formatDuration];
     if (message && _delegate && [_delegate respondsToSelector:@selector(inputController:didSendMessage:)]) {
